% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_test.R
\name{stat_test}
\alias{stat_test}
\title{Perform Two Sample T-Test and Chi-Square Test on Data}
\usage{
stat_test(data, target_var)
}
\arguments{
\item{data}{Data frame object}

\item{target_var}{Target variable name of data}
}
\description{
The function performs two sample t-test and chi-square test and exports the output of each test in a text file.
}
\examples{
stat_test(Boston)

}
